# api/src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: Product Order Service API
  version: 1.0.0
  description: E-commerce platform API for product and order management
servers:
  - url: http://localhost:8080/product-order-service
    description: Local development
  - url: http://13.202.157.59:8080/product-order-service
    description: AWS EC2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
...
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        sku:
          type: string
        isActive:
          type: boolean
        category:
          $ref: '#/components/schemas/CategoryResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      required:
        - name
        - price
        - stockQuantity
        - sku
        - categoryId
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        sku:
          type: string
        categoryId:
          type: integer
          format: int64
        isActive:
          type: boolean
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        isActive:
          type: boolean
        categoryId:
          type: integer
          format: int64
    UpdateStockRequest:
      type: object
      required:
        - stockQuantity
      properties:
        stockQuantity:
          type: integer

    # Category
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean

    # Order
...
paths:
...
  /api/v1/products:
    post:
      tags: [Products]
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

    get:
      tags: [Products]
      operationId: getAllProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/products/{productId}:
    get:
      tags: [Products]
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

    put:
      tags: [Products]
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

    delete:
      tags: [Products]
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/products/sku/{sku}:
    get:
      tags: [Products]
      operationId: getProductBySku
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
  /api/v1/products/search:
    get:
      tags: [Products]
      operationId: searchProducts
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /api/v1/products/category/{categoryId}:
    get:
      tags: [Products]
      operationId: getProductsByCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /api/v1/products/price-range:
    get:
      tags: [Products]
      operationId: getProductsByPriceRange
      parameters:
        - name: minPrice
          in: query
          required: true
          schema:
            type: number
        - name: maxPrice
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /api/v1/products/search/advanced:
    get:
      tags: [Products]
      operationId: advancedSearchProducts
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /api/v1/products/low-stock:
    get:
      tags: [Products]
      operationId: getLowStockProducts
      security:
        - bearerAuth: []
      parameters:
        - name: threshold
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
  /api/v1/products/{productId}/stock:
    patch:
      tags: [Products]
      operationId: updateProductStock
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

  # Orders
...